Refactor boot_set.py into submodule for clarity
